src/lumi-bot/pyproject.toml
[project]
name = "lumi-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.0",
    "aiohttp-cors>=0.8.1",
    "dotenv>=0.9.9",
    "librosa>=0.11.0",
    "noisereduce>=3.0.3",
    "numpy>=2.2.6",
    "openai-whisper>=20240930",
    "py-cord[speed,voice]>=2.6.1",
    "python-socketio>=5.13.0",
]

[project.optional-dependencies]
cpu = ["torch>=2.7.0", "torchaudio>=2.7.0"]
rocm = [
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "pytorch-triton-rocm>=3.3.0 ; sys_platform == 'linux'",
]
cu128 = ["torch>=2.7.0", "torchaudio>=2.7.0"]

[tool.uv]
constraint-dependencies = ["torch>=2.7.0", "torchaudio>=2.7.0"]
conflicts = [[{ extra = "cpu" }, { extra = "cu128" }, { extra = "rocm" }]]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-rocm", extra = "rocm" },
    { index = "pytorch-cu128", extra = "cu128" },
]
torchaudio = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-rocm", extra = "rocm" },
    { index = "pytorch-cu128", extra = "cu128" },
]
# ROCm support relies on `pytorch-triton-rocm`, which should also be installed from the PyTorch index
pytorch-triton-rocm = [
    { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]
py-cord = { git = "https://github.com/Pycord-Development/pycord", branch = "master" }

[tool.uv.pip]
extra = ["rocm"]

[dependency-groups]
dev = ["black>=25.1.0", "ipykernel>=6.29.5", "mypy>=1.15.0", "pylint>=3.3.7"]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.3"
explicit = true
